refactor: トランスクリプトの引き方を共通サービスへ切り出す
SessionModelService が持っていた「作業フォルダ -> ~/.claude/projects の記録」の
解決と jsonl の末尾読みを ClaudeTranscriptLocator / TranscriptTail へ移す。
コンテキスト量バッジが同じ引き方をするため、先に共通化する。

挙動は変えていない（移動のみ）。
